requirements.txt
aiohttp==3.7.4.post0
asyncpg==0.24.0
discord.py==1.7.3
emojis==0.6.0
humanize==3.12.0
jishaku==2.3.2
loguru==0.5.3
matplotlib==3.4.3
psutil==5.8.0
pyparsing==3.0.0
python_dateutil==2.8.2
tekore==4.1.0
vacefron.py==1.6.3
